定数はヘッダーに書かない opeator<<=,>>-,|=の導入 IsZeroはoperator==を利用 usingの見直し include不足の解消

+++ Iona/types.hpp
@@ -8,8 +8,7 @@
 #pragma once
 #include<cstdint>
 
-using std::size_t;
-
+using std::uint8_t;
 /**
 * @enum Turn
 * Information of turn player
